Created car agent main class and subclasses
Car agent subclass ok for navigation

Need to fix track1 subclass
