--- Assets/Scripts/car_agents/car_agent_discrete1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 90e862f4751be1c44afd767495b1a02d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
